feat: rename GildedRoseTest to GildedRoseAgedBrieTests and refactor related tests
